pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "leoflow-store"
version = "0.1.0"
description = "Local workflow store and CLI for LeoFlow."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
lf = "leoflow_store.cli:main"

[tool.setuptools.packages.find]
include = ["leoflow_store*"]

[tool.setuptools.package-data]
leoflow_store = ["templates/*/*.tpl"]
